Remove comments
Not setting the tag field of the target variable would result in
target.tag->name having a strangely predictable value - either
'XDG_SESSION_CLASS=user' or another similar string for another XDG
variable. Expected a segfault due to a garbage address, but gcc seems to
do something that makes this not the case. gcc on another machine as
clang on this machine gives a segfault as expected.

In conclusion, the curious behavior was caused by a compiler
optimization.