Remove comments and rename variable
In points_at_callback, we don't have to worry about the tag pointing to
something that's not a commit because we will always be querying for a
tag that matches a certain commit id, so if the tag doesn't target a
commit, it won't match the id.

Rename this_tag to cur_tag to better reflect the function's purpose and
remove any possible similarity this could have to JavaScript code.