Allow pre-existing release message
If .hubrelease_message exists, use it as-is. Else, if the release is a
draft, write the tag message to .hubrelease_message and continue. Else,
edit .hubrelease_message with the user's preferred editor.

Remove .hubrelease_message after posting the release.

Log a warning if opening an editor when stdout is not connected to a
terminal.